backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
scikit-learn
numpy
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
pdfplumber
openpyxl
python-dotenv
